V1.5.1 修复创建音乐列表报错问题 [author]        wm [reason]   创建列表时，往数据插入数据时马上获取该数据的id，数据库没有插入完成，无法获取 [solution] 插入数据时，通过last_insert_rowid()函数来获取刚插入的自动递增id，而不是根据列表名来获取

+++ MediaPlayProject/app/src/main/java/com/example/mediaplayproject/bean/MusicListBean.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public class MusicListBean {
     private String listName;
-    private int listId;
+    private long listId;
     /**
     * 注意：这里的size并不能同步的获取list的音乐数量需要手动的获取list.size()
     * */
@@ -33,11 +33,11 @@ public class MusicListBean {
         this.listName = listName;
     }
 
-    public int getListId() {
+    public long getListId() {
         return listId;
     }
 
-    public void setListId(int listId) {
+    public void setListId(long listId) {
         this.listId = listId;
     }
